Bug 354661 - time-admin fails to lookup "Europe/Helsinki" as time zone value Bug 354661 -  time-admin fails to lookup "Europe/Helsinki" as time zone value (Debian/Ubuntu)
The problem is that there are two zoneinfo's that are identical, but system-tools-backend just picks the first one in /usr/share/zoneinfo/zone.tab that matches.

According to http://wiki.debian.org/TimeZoneChanges, Etch and later use /etc/timezone as the preferred location for that info.  Etch is Debian 4.0.According to https://bugs.edge.launchpad.net/ubuntu/+source/tzdata/+bug/209429/comments/16, Ubuntu Edgy and later do the same thing.
So my patch changes the backend for debian-4.0 (Etch) and ubuntu-7.04 (Feisty) and up.