Use PAM to set password
Instead of getting the password encrypted by the client, get it in plain text, and set it via PAM. This behavior is standard on all supported platforms, which allows for code simplification, and avoids bugs when new password encryption algorithms appear.

Never return the encrypted password on the bus. Instead, PAM should always be used to change it, possibly authenticating when changing own password.