Enable ServicesList filtering to hide critical services
Use the list of "forbidden" services that was present in ServicesList.pm to filter services we send to the clients. This list includes every service that should not be easily disabled since they could hurt the system. With this change, clients can show all services to the user, instead of limiting to those they know about.

Add many services from Ubuntu, which are likely to exist on other distributions. In particular, we now hide gdm and dbus since disabling those is rarely needed and can be dangerous - it would be equivalent to shooting us in the arm, anyway.