Improve error checking against malformed /etc/group entries
When a group line contained an extra ':', the group struct contained 5 items instead of 4, leading D-Bus bindings to die unexpectedly. Clients did not receive any group struct, and could remove all groups on the system. This bug was mitigated by the move to individual commits, but this change will avoid dying for most cases. UsersConfig doesn't seem to suffer from the same problem.

https://bugs.launchpad.net/ubuntu/+source/gnome-system-tools/+bug/160862