Fix wrong int size when getting UID in dispatch_self_config()
dbus_message_get_args() with type DBUS_TYPE_UINT32 requires passing a reference to a guint32. The gulong we were using was 64 bits long on amd64 platforms, meaning the higher half of the UID was garbage. This fixes a bug preventing people from changing their own settings, which occurred because the passed UID was rejected.