Identify users by login, not UID
Login is considered to be the fixed way of identifying users and groups. This allows changing their UID. Changing a user's login is not possible since we would consider it as a new user. This is how things were supposed to work (see the gnome-system-tools).

Also see commit 34e6995, which fixed the same mistake for groups.